perf: complete operator runtime refactor (#18)
Collapse the operator-runtime performance work into one coherent change.

This wires semantic benchmark gates and local CI coverage, moves the hot scan/join/aggregate/sort paths onto typed runtime structures, hardens spill and reclaimer ownership, and retires legacy compatibility seams around row formats, specs, lowering, and scheduler work units.

It also locks in the review follow-up guardrails for aggregate DISTINCT/ORDERED replay, classic IEJoin cursor emission, runtime filters, profile output, and operator-runtime naming.

+++ crates/storage/src/row/pinned.rs
@@ -323,7 +323,7 @@ impl PinnedRow<'_, '_> {
     }
 
     #[inline]
-    pub fn ordinal(&self) -> u32 {
+    pub fn ordinal(&self) -> u64 {
         self.rows.logical_rows[self.logical_index].ordinal
     }
